pyproject.toml
[project]
name = "markdown-table-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "google-genai>=1.50.1",
    "numpy>=2.3.4",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pymupdf>=1.26.6",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
